fix clippy nits and move serialize focused tests to serializer file
